Created MoonElevationModel.java to import local elevation data
